Socket packet test
